pyproject.toml
[tool.poetry]
name = "quickex-auth"
version = "0.1.0"
description = ""
authors = ["SSomov <goofma@gmail.com>"]
readme = "README.md"
packages = [{include = "quickex_auth"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
